fix: layour and file directory of user settings
