+++ src/plugins/i18n/locales/en.json
@@ -116,7 +116,7 @@
       },
       "title": "Settings",
       "users": {
-        "description": "Manage your users",
+        "description": "View and manage all your users",
         "title": "User settings"
       }
     },
